oiblz

+++ src/account.rs
@@ -38,18 +38,18 @@ impl Account {
             Ok(account) => {
                 for balance in account.balances {
                     if balance.asset == asset {
-                        return Ok(balance)     
+                        return Ok(balance)
                     }
                 }
-                bail!(format!("Asset not found")); 
+                bail!(format!("Asset not found"));
             },
             Err(e) => return Err(e),
-        } 
-    }    
+        }
+    }
 
     // Current open orders for ONE symbol
     pub fn get_open_orders(&self, symbol: String) -> Result<(Vec<Order>)> {
-        
+
         let mut parameters: BTreeMap<String, String> = BTreeMap::new();
         parameters.insert("symbol".into(), symbol);
 
@@ -62,7 +62,7 @@ impl Account {
 
     // Check an order's status
     pub fn order_status(&self, symbol: String, order_id: u32) -> Result<(Order)> {
-        
+
         let mut parameters: BTreeMap<String, String> = BTreeMap::new();
         parameters.insert("symbol".into(), symbol);
         parameters.insert("orderId".into(), order_id.to_string());
@@ -72,11 +72,11 @@ impl Account {
         let order: Order = from_str(data.as_str()).unwrap();
 
         Ok(order)
-    }    
+    }
 
     // Place a LIMIT order - BUY
-    pub fn limit_buy(&self, symbol: String, qty: u32, price: f64) -> Result<(Transaction)> {
-        
+    pub fn limit_buy(&self, symbol: String, qty: f64, price: f64) -> Result<(Transaction)> {
+
         let order = self.build_order(symbol, qty, price, ORDER_SIDE_BUY, ORDER_TYPE_LIMIT, TIME_IN_FORCE_GTC);
         let request = build_signed_request(order, self.recv_window);
         let data = self.client.post_signed(API_V3_ORDER, request)?;
@@ -86,19 +86,19 @@ impl Account {
     }
 
     // Place a LIMIT order - SELL
-    pub fn limit_sell(&self, symbol: String, qty: u32, price: f64) -> Result<(Transaction)> {
-        
+    pub fn limit_sell(&self, symbol: String, qty: f64, price: f64) -> Result<(Transaction)> {
+
         let order = self.build_order(symbol, qty, price, ORDER_SIDE_SELL, ORDER_TYPE_LIMIT, TIME_IN_FORCE_GTC);
         let request = build_signed_request(order, self.recv_window);
         let data = self.client.post_signed(API_V3_ORDER, request)?;
         let transaction: Transaction = from_str(data.as_str()).unwrap();
 
         Ok(transaction)
-    }    
+    }
 
     // Place a MARKET order - BUY
-    pub fn market_buy(&self, symbol: String, qty: u32) -> Result<(Transaction)> {
-      
+    pub fn market_buy(&self, symbol: String, qty: f64) -> Result<(Transaction)> {
+
         let order = self.build_order(symbol, qty, 0.0, ORDER_SIDE_BUY, ORDER_TYPE_MARKET, TIME_IN_FORCE_GTC);
         let request = build_signed_request(order, self.recv_window);
         let data = self.client.post_signed(API_V3_ORDER, request)?;
@@ -108,8 +108,8 @@ impl Account {
     }
 
     // Place a MARKET order - SELL
-    pub fn market_sell(&self, symbol: String, qty: u32) -> Result<(Transaction)> {
-        
+    pub fn market_sell(&self, symbol: String, qty: f64) -> Result<(Transaction)> {
+
         let order = self.build_order(symbol, qty, 0.0, ORDER_SIDE_SELL, ORDER_TYPE_MARKET, TIME_IN_FORCE_GTC);
         let request = build_signed_request(order, self.recv_window);
         let data = self.client.post_signed(API_V3_ORDER, request)?;
@@ -120,7 +120,7 @@ impl Account {
 
     // Check an order's status
     pub fn cancel_order(&self, symbol: String, order_id: u32) -> Result<(OrderCanceled)> {
-        
+
         let mut parameters: BTreeMap<String, String> = BTreeMap::new();
         parameters.insert("symbol".into(), symbol);
         parameters.insert("orderId".into(), order_id.to_string());
@@ -130,11 +130,11 @@ impl Account {
         let order_canceled: OrderCanceled = from_str(data.as_str()).unwrap();
 
         Ok(order_canceled)
-    } 
+    }
 
     // Trade history
     pub fn trade_history(&self, symbol: String) -> Result<(Vec<TradeHistory>)> {
-        
+
         let mut parameters: BTreeMap<String, String> = BTreeMap::new();
         parameters.insert("symbol".into(), symbol);
 
@@ -143,23 +143,23 @@ impl Account {
         let trade_history: Vec<TradeHistory> = from_str(data.as_str()).unwrap();
 
         Ok(trade_history)
-    }     
+    }
 
-    fn build_order(&self, symbol: String, qty: u32, price: f64, 
+    fn build_order(&self, symbol: String, qty: f64, price: f64,
                    order_side: &str, order_type: &str, time_in_force: &str) ->  BTreeMap<String, String> {
-        
+
         let mut order: BTreeMap<String, String> = BTreeMap::new();
 
         order.insert("symbol".into(), symbol);
         order.insert("side".into(), order_side.to_string());
-        order.insert("type".into(), order_type.to_string());       
+        order.insert("type".into(), order_type.to_string());
         order.insert("quantity".into(), qty.to_string());
 
         if price != 0.0 {
             order.insert("price".into(), price.to_string());
-            order.insert("timeInForce".into(), time_in_force.to_string()); 
+            order.insert("timeInForce".into(), time_in_force.to_string());
         }
-        
+
         order
     }
 }
